requirements.txt
"""
PyMuPDF==1.23.26
google-generativeai==0.3.2
faiss-cpu==1.7.4
scikit-learn==1.3.2
numpy==1.24.3
python-dotenv==1.0.0
"""